+++ CHANGELOG.md
@@ -6,6 +6,11 @@ v1.4.1
 * Added support two-phase commit, using the new transaction prepare and getId methods.
 * Added a parallel external mergesort utility.
 
+v1.4.0.2 (2018-01-21)
+--------
+* Fix bug when deleting entries from the largest allowed page size, 65536 bytes. Under the right
+  conditions, the internal search vector pointer would overflow the 16-bit range.
+
 v1.4.0.1 (2018-01-01)
 --------
 * Added a RAFT-based replication system.